+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,18 +1,18 @@
-package kitchenpos.ui;
+package kitchenpos.product.ui;
 
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
+import java.net.URI;
+import java.util.List;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
-import java.util.List;
-
 @RestController
 public class ProductRestController {
+
     private final ProductService productService;
 
     public ProductRestController(final ProductService productService) {
@@ -23,15 +23,11 @@ public class ProductRestController {
     public ResponseEntity<Product> create(@RequestBody final Product product) {
         final Product created = productService.create(product);
         final URI uri = URI.create("/api/products/" + created.getId());
-        return ResponseEntity.created(uri)
-                .body(created)
-                ;
+        return ResponseEntity.created(uri).body(created);
     }
 
     @GetMapping("/api/products")
     public ResponseEntity<List<Product>> list() {
-        return ResponseEntity.ok()
-                .body(productService.list())
-                ;
+        return ResponseEntity.ok().body(productService.list());
     }
 }
